RMX1805: Update SurfaceFlinger pin
This change replaces the '/system/lib64/libsurfaceflinger.so' pin, as the file was
removed in# Please enter the commit message for your changes. Lines starting
ag/12524602. The updated pin relies on# with '#' will be ignored, and an empty
message aborts the commit.
go/aog/1552085, as system_service needs read access to SurfaceFlinger.